feat: DishService now has correct delete logic (from requirements).
fix: init.sql cascade delete.

+++ src/main/java/com/example/demo/services/DishService.java
@@ -1,9 +1,11 @@
 package com.example.demo.services;
 
 import com.example.demo.entities.Dish;
+import com.example.demo.exceptions.ItemInUseException;
 import com.example.demo.exceptions.ItemNotFoundException;
 import com.example.demo.repository.DishRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,9 +24,24 @@ public class DishService {
 		return dishRepository.findById(id);
 	}
 
+	/**
+	 * Removes Dish from the menu only if it was not ordered.
+	 * the `DishIngredient` table has entries removed as well.
+	 *
+	 * @param id Dish to remove
+	 */
 	public void deleteById(Long id) {
 		if (!dishRepository.existsById(id))
 			throw new ItemNotFoundException(Dish.class, id);
+		Specification<Dish> dishHasOrders = (root, query, cb) -> {
+			return cb.and(
+				cb.equal(root.get(Dish.id_), id), // filter by Dish ID
+				cb.isNotEmpty(root.get(Dish.ordersDishes_)) // Check for orders in that dish
+			);
+		};
+		//check if predicate is true
+		if (dishRepository.exists(dishHasOrders))
+			throw new ItemInUseException(Dish.class, id);
 		dishRepository.deleteById(id);
 	}
 }
